+++ nequip/data/dataset/npz_dataset.py
@@ -4,13 +4,13 @@ import torch
 
 from .. import AtomicDataDict
 from ..dict import from_dict
-from ._base_datasets import AtomicDataset
+from .base_datasets import AtomicDataset
 
 from typing import Union, Dict, List, Callable
 
 
 class NPZDataset(AtomicDataset):
-    """``AtomicDataset`` that loads data from an NPZ file following `sGDML <https://www.sgdml.org/#datasets>`_ conventions.
+    """:class:`~nequip.data.dataset.AtomicDataset` that loads data from an NPZ file following `sGDML <https://www.sgdml.org/#datasets>`_ conventions.
     It is also compatible with other datasets such as rMD-17, with a change in ``key_mapping``
     (the default ``key_mapping`` is set to be compatible with sGDML datasets).