+++ nequip/nn/graph_model.py
@@ -14,12 +14,12 @@ NUM_TYPES_KEY: Final[str] = "num_types"
 MODEL_DTYPE_KEY: Final[str] = "model_dtype"
 
 
-def _model_metadata_from_config(model_config: Dict[str, str]) -> Dict[str, Any]:
+def _model_metadata_from_config(model_config: Dict[str, str]) -> Dict[str, str]:
     model_metadata_dict = {}
     # manually process everything
     model_metadata_dict[MODEL_DTYPE_KEY] = model_config[MODEL_DTYPE_KEY]
     model_metadata_dict[TYPE_NAMES_KEY] = " ".join(model_config[TYPE_NAMES_KEY])
-    model_metadata_dict[NUM_TYPES_KEY] = len(model_config[TYPE_NAMES_KEY])
+    model_metadata_dict[NUM_TYPES_KEY] = str(len(model_config[TYPE_NAMES_KEY]))
     model_metadata_dict[R_MAX_KEY] = str(model_config[R_MAX_KEY])
 
     if PER_EDGE_TYPE_CUTOFF_KEY in model_config:
@@ -50,6 +50,8 @@ class GraphModel(GraphModuleMixin, torch.nn.Module):
     is_graph_model: Final[bool] = True
     # ^ to identify `GraphModel` types from `nequip-package`d models (see https://pytorch.org/docs/stable/package.html#torch-package-sharp-edges)
 
+    _metadata: Dict[str, str]
+
     def __init__(
         self,
         model: GraphModuleMixin,
@@ -61,6 +63,7 @@ class GraphModel(GraphModuleMixin, torch.nn.Module):
             # Things that always make sense as inputs:
             AtomicDataDict.POSITIONS_KEY: "1o",
             AtomicDataDict.EDGE_INDEX_KEY: None,
+            AtomicDataDict.EDGE_TRANSPOSE_PERM_KEY: None,
             AtomicDataDict.EDGE_CELL_SHIFT_KEY: None,
             AtomicDataDict.CELL_KEY: "1o",  # 3 of them, but still
             AtomicDataDict.BATCH_KEY: None,
@@ -80,15 +83,24 @@ class GraphModel(GraphModuleMixin, torch.nn.Module):
 
         # the following logic is for backward compatibility and to simplify unittests
         self.model_dtype = torch.get_default_dtype()
-        self.metadata = {}
+        self._metadata = {}
         self.type_names = []
         if model_config is not None:
-            self.metadata = _model_metadata_from_config(model_config)
-            self.type_names = self.metadata[TYPE_NAMES_KEY].split(" ")
+            self._metadata = _model_metadata_from_config(model_config)
+            self.type_names = self._metadata[TYPE_NAMES_KEY].split(" ")
             model_dtype = {"float32": torch.float32, "float64": torch.float64}[
-                self.metadata[MODEL_DTYPE_KEY]
+                self._metadata[MODEL_DTYPE_KEY]
             ]
             assert self.model_dtype == model_dtype
+
+    @property
+    @torch.jit.unused
+    def metadata(self) -> Dict[str, str]:
+        # Note that this is a property so that the metadata can depend on the _current_ state
+        # of the model, and not just what happened at initialization.
+        # TODO: make other metadata keys dynamic rather than pre-set in _metadata?
+        out = self._metadata.copy()
+        return out
 
     def forward(self, data: AtomicDataDict.Type) -> AtomicDataDict.Type:
         # restrict the input data to allowed keys to prevent the model from directly using the dict from the outside,